Split the param prompts into two test files
