+++ tests/cypress/e2e/param-prompts-python.cy.ts
@@ -1,0 +1,24 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+require("cypress-terminal-report/src/installLogsCollector")();
+import "@testing-library/cypress/add-commands";
+import failOnConsoleError from "cypress-fail-on-console-error";
+failOnConsoleError();
+// For the beforeEach part:
+import "../support/param-prompt-support";
+import {testRawFuncs} from "../support/param-prompt-support";
+
+
+describe("Parameter prompts Python only", () => {
+    // Each item is a triple: the module, the function name within the module, the list of param names
+    const rawFuncs : [string | null | [string, string], string, string[]][] = [];
+    if (Cypress.env("mode") !== "microbit") {
+        rawFuncs.push(["urllib.request", "urlopen", ["url"]]);
+        rawFuncs.push(["turtle", "Turtle", []]);
+        rawFuncs.push(["datetime", "date.fromtimestamp", ["timestamp"]]);
+        rawFuncs.push(["strype.graphics", "load_image", ["name"]]);
+        rawFuncs.push(["strype.graphics", "Actor", ["image_or_filename"]]);
+        rawFuncs.push([["http://localhost:8089/test-library/", "mediacomp"], "makePicture", ["path"]]);
+        rawFuncs.push([["http://localhost:8089/test-library/", "mediacomp"], "Pixel", ["picture", "x", "y", "index"]]);
+    }
+    testRawFuncs(rawFuncs);
+});
